requirements.txt
qrcode[pil]
reportlab
mysql-connector-python # Comentado para usar SQLite por padrão

